% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElements.R
\name{addText}
\alias{addText}
\title{Add text to plot}
\usage{
addText(text)
}
\arguments{
\item{text}{Character.  Text to be added to the plot.}
}
\description{
Utility function for adding general text as meta-information
}
