% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addUserPath}
\alias{addUserPath}
\title{Add path to plot}
\usage{
addUserPath(userID = TRUE, path = TRUE, fontsize = 12)
}
\arguments{
\item{userID}{Logical, should the user ID be included with the path}

\item{path}{Logical or character, should the path to the current working directory,
or a provided path, be included.}

\item{fontsize}{The font size in pt}
}
\description{
Add path to plot
}
