% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_lines}
\alias{count_lines}
\title{Count string lines}
\usage{
count_lines(string)
}
\arguments{
\item{string}{a character vector of length 1.}
}
\description{
Count string lines based on character returns
}
