% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countMeta}
\alias{countMeta}
\title{Count meta lines}
\usage{
countMeta(meta)
}
\arguments{
\item{meta}{Lines of meta information for a single meta data item.  The function counts lines based on character returns, \code{"\n"} and the number of elements in the vector.}
}
\value{
numeric. The total number of lines of meta information.
}
\description{
Function to count the number of lines of meta information
}
