% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countMetaMulti}
\alias{countMetaMulti}
\title{Count meta lines}
\usage{
countMetaMulti(...)
}
\arguments{
\item{...}{\code{\link{countMeta}} Multiple metadata items to be counted by `countMeta`}
}
\value{
numeric. The total number of lines of meta information.
}
\description{
Function to count the number of lines of meta information
}
