% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createUserPath}
\alias{createUserPath}
\title{Create User & Path String}
\usage{
createUserPath(userID = TRUE, path = TRUE)
}
\arguments{
\item{userID}{logical. Should the user ID be included with the path}

\item{path}{logical or character. If logical, should current working directory 
be added as the path. If character, the path to be included.}
}
\description{
Create User & Path String
}
