% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGGTitle.R
\name{deleteGGTitle}
\alias{deleteGGTitle}
\title{Remove the title from a ggplot object}
\usage{
deleteGGTitle(object)
}
\arguments{
\item{object}{A ggplot or lattice object}
}
\value{
The original object, minus the title
}
\description{
Remove the title from a ggplot object
}

