% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGGTitle.R
\name{extractGGTitle}
\alias{extractGGTitle}
\title{Extract the title from a ggplot object}
\usage{
extractGGTitle(object)
}
\arguments{
\item{object}{A ggplot or lattice object}
}
\value{
Character tile or NULL if it doesn't exist
}
\description{
Extract the title from a ggplot object
}
