% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTag.R
\name{ggTag}
\alias{ggTag}
\title{Tag a ggplot2 graphic with meta information}
\usage{
ggTag(object, raster = FALSE, extractTitle = FALSE, title = NULL,
  meta = list(), fontsize = 12, theme = NULL, inherit_size = FALSE)
}
\arguments{
\item{object}{A ggplot or lattice object}

\item{raster}{Logical. Defaults to FALSE. Set to TRUE if input is a matrix to raster.}

\item{extractTitle}{Logical. Defaults to FALSE. Extract the title from the graph and use as plot title.}

\item{title}{Character.You own title.  Overridden if extractTitle is TRUE.}

\item{meta}{List containing meta information to include in the 4 corners of the plot: top_left, top_right, bottom_left, bottom_right}

\item{fontsize}{The font size in pt.}

\item{theme}{An optional ggplot2 theme to use.  Only applies to the current plot, i.e. the main theme is not updated.}

\item{inherit_size}{Logical.  If `TRUE` the value of fontsize is passed through to the selected ggplot2 theme.}
}
\description{
Tag a ggplot2 or lattice (or any grid) object with meta information by nesting it within a grid framework
}
\examples{
{
library(grid)
library(ggplot2)
library(mangoTraining)
# Layout used:
grid.show.layout(
grid.layout(nrow = 3, ncol = 3,
	heights = unit(c(4, 1, 3), c("lines", "null", "lines")),
	widths = unit(c(.25, 1, .25), c("inches", "null", "inches"))
)
)

# Create ggplot2 graphic
myPlot <- qplot(Weight, Height, data = demoData, facets = Sex ~ Smokes,
  main = "Scatter Plot of Height against Weight\\nby Sex and Smoking Status\\n",
  xlab = "\\nWeight",
  ylab = "Height\\n"
)
myPlot <- myPlot + theme_bw()
myPlot <- myPlot +
  theme(strip.text.y = element_text(),
        strip.background = element_rect(fill = NA, linetype = 0),
        panel.grid.minor = element_line(colour = NA),
        panel.grid.major = element_line(colour = NA)
  )
myPlot


ggTag(myPlot, 
      meta = list(top_left = "Protocol: 123456\\nPopulation: Intent-to-Treat",
                  top_right = "Page 1 of 1",    
                  bottom_left = paste(Sys.getenv("USERNAME"), getwd()), 
                  bottom_right = date()))
      
# Reduce font size
ggTag(myPlot, 
      meta = list(top_left = "Protocol: 123456\\nPopulation: Intent-to-Treat",
                  top_right = "Page 1 of 1",    
                  bottom_left = paste(Sys.getenv("USERNAME"), getwd()), 
                  bottom_right = date()), 
                  fontsize = 8)
                  
 # Change theme          
ggTag(myPlot, 
      meta = list(top_left = "Protocol: 123456\\nPopulation: Intent-to-Treat",
                  top_right = "Page 1 of 1",    
                  bottom_left = paste(Sys.getenv("USERNAME"), getwd()), 
                  bottom_right = date()),
                  theme = theme_grey())       
}
}
