% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmaTag.R
\name{pharmaTag}
\alias{pharmaTag}
\title{Tag a ggplot2 graphic with meta information, specific to phara requirements}
\usage{
pharmaTag(object, protocol = NA, population = NA, page = 1, pages = 1,
  ...)
}
\arguments{
\item{object}{A ggplot or lattice object}

\item{protocol}{What is the protocol number?}

\item{population}{What is the population?}

\item{page}{What is the page?}

\item{pages}{How many pages are there?}

\item{...}{Arguments to pass to `ggTag`}
}
\description{
A wrapper for `ggTag`, specific to the pharma industry
}
