% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findJustification.R
\name{position}
\alias{position}
\title{Find the x, y and justification values}
\usage{
position(topLines = 2)
}
\arguments{
\item{topLines}{The number of lines of meta-data at the top of the plot}
}
\description{
Based on the viewport that the element is being pushed into define the x, y and 
justification values.
}
