% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addDateTime}
\alias{addDateTime}
\title{Add time to plot}
\usage{
addDateTime(dateFormat = "\%d\%b\%Y \%H:\%M")
}
\arguments{
\item{dateFormat}{Character.  R date format to use for the date.}
}
\description{
Add time to plot
}
