% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElements.R
\name{addUserPath}
\alias{addUserPath}
\title{Add path to plot}
\usage{
addUserPath(userID = TRUE, path = TRUE)
}
\arguments{
\item{userID}{Logical. Should the username be included}

\item{path}{Logical or character. If logical, should the working directory be
added to the system path. If character, the system path to be added.}
}
\description{
Utility function for adding username and system path
}
