% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTag.R
\name{ggTag}
\alias{ggTag}
\title{Tag a ggplot2 graphic with meta information}
\usage{
ggTag(object, meta1 = "", meta2 = "", date = TRUE, username = TRUE,
  path = TRUE)
}
\arguments{
\item{object}{A ggplot or lattice object}

\item{meta1}{A line of meta information}

\item{meta2}{A second line of meta information to appear below the first}

\item{date}{Logical. Defaults to TRUE.}

\item{username}{Logical. Defaults to TRUE.}

\item{path}{Logical. Defaults to TRUE.}
}
\description{
Tag a ggplot2 or lattice (or any grid) object with meta information by nesting it within a grid framework
}
\examples{
{
library(grid)
library(ggplot2)
library(mangoTraining)
# Layout used:
grid.show.layout(
grid.layout(nrow = 3, ncol = 3,
	heights = unit(c(4, 1, 3), c("lines", "null", "lines")),
	widths = unit(c(.25, 1, .25), c("inches", "null", "inches"))
)
)

# Create ggplot2 graphic
myPlot <- qplot(Weight, Height, data = demoData, facets = Sex ~ Smokes,
  main = "Scatter Plot of Height against Weight\\nby Sex and Smoking Status\\n",
  xlab = "\\nWeight",
  ylab = "Height\\n"
)
myPlot <- myPlot + theme_bw() 
myPlot <- myPlot + 
  theme(strip.text.y = element_text(), 
        strip.background = element_rect(fill = NA, linetype = 0),
        panel.grid.minor = element_line(colour = NA),
        panel.grid.major = element_line(colour = NA)
  )
myPlot


ggTag(myPlot, meta1 = "Protocol: 123456", meta2 = "Study: 123456", 
      date = TRUE, username = TRUE, path = FALSE)
}
}

