% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTag.R
\name{ggTag}
\alias{ggTag}
\title{#' Tag a ggplot2 graphic with meta information}
\usage{
ggTag(object, topLeft = NULL, topRight = NULL, bottomLeft = NULL,
  bottomRight = NULL)
}
\arguments{
\item{object}{A ggplot or lattice object}

\item{topLeft}{Grid object}

\item{topRight}{Grid object}

\item{bottomLeft}{Grid object}

\item{bottomRight}{Grid object}
}
\description{
Tag a ggplot2 or lattice (or any grid) object with meta information by 
nesting it within a grid framework
}
\examples{
{
library(mangoTraining)
library(ggplot2)
myPlot <- qplot(Weight, Height, data = demoData, facets = Sex ~ Smokes,
  main = "Scatter Plot of Height against Weight\\nby Sex and Smoking Status\\n",
  xlab = "\\nWeight",
  ylab = "Height\\n"
)

ggTag(myPlot, 
  bottomLeft = addUserPath(),
  bottomRight = addDateTime(),
  topLeft = addTest("Project ABC")
 )

}
}
