% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmaTag.R
\name{pharmaTag}
\alias{pharmaTag}
\title{Tag a ggplot2 graphic with meta information, specific to phara requirements}
\usage{
pharmaTag(object, protocol = NA, population = NA, page = 1, pages = 1,
  date = TRUE, username = TRUE, path = TRUE,
  date_format = "\%d\%b\%Y \%H:\%M", ...)
}
\arguments{
\item{object}{A ggplot or lattice object}

\item{protocol}{What is the protocol number?}

\item{population}{What is the population?}

\item{page}{What is the page?}

\item{pages}{How many pages are there?}

\item{date}{Logical. Defaults to TRUE.}

\item{username}{Logical. Defaults to TRUE.}

\item{path}{Logical. Defaults to TRUE.}

\item{date_format}{Character.  R date format to use for the date.}

\item{...}{Arguments to pass to `ggTag`}
}
\description{
A wrapper for `ggTag`, specific to the pharma industry
}
\examples{
{
library(grid)
library(ggplot2)
library(mangoTraining)
# Layout used:

# Create ggplot2 graphic
myPlot <- qplot(Weight, Height, data = demoData, facets = Sex ~ Smokes,
  main = "Scatter Plot of Height against Weight\\nby Sex and Smoking Status\\n",
  xlab = "\\nWeight",
  ylab = "Height\\n"
)
myPlot <- myPlot + theme_bw()
myPlot <- myPlot +
  theme(strip.text.y = element_text(),
        strip.background = element_rect(fill = NA, linetype = 0),
        panel.grid.minor = element_line(colour = NA),
        panel.grid.major = element_line(colour = NA)
  )


pharmaTag(myPlot, fontsize = 8)
      
}
}
